# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GreenLync Event-Driven API Gateway Boilerplate",
        "title": "GreenLync API Gateway",
        "termsOfService": "https://greenlync.com",
        "contact": {
            "name": "API Support",
            "email": "zeelrupapara@gmail.com"
        },
        "license": {
            "name": "GreenLync",
            "url": "https://greenlync.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/api/v1/configs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Configs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "operationId": "GetAllConfigs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Config"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/groups/{group_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Configs by Group ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "operationId": "GetConfigsBelongToGroup",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Config"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/{config_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Config by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "operationId": "GetConfig",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Config ID",
                        "name": "config_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "http.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Http status Code",
                    "type": "integer"
                },
                "data": {
                    "description": "if the request were successful the data will be saved here"
                },
                "error": {
                    "description": "Generic General Error Message defined in the system",
                    "type": "string"
                },
                "message": {
                    "description": "More detailed error message indicates why the request was unsuccessful",
                    "type": "string"
                },
                "success": {
                    "description": "Response flag indicates whether the HTTP request was successful or not",
                    "type": "boolean"
                }
            }
        },
...
        "model.Config": {
            "type": "object",
            "properties": {
                "config_group": {
                    "$ref": "#/definitions/model.ConfigGroup"
                },
                "config_group_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_public": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "record_type": {
                    "$ref": "#/definitions/model.RecordType"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "value_type": {
                    "$ref": "#/definitions/model.ValueType"
                }
            }
        },
        "model.ConfigGroup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_system": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.RecordType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "RecordType_Seed",
                "RecordType_User",
                "RecordType_System",
                "RecordType_Import"
            ]
        },
...
        "model.ValueType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "ValueType_String",
                "ValueType_Integer",
                "ValueType_Boolean",
                "ValueType_Float",
                "ValueType_JSON"
            ]
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Type \"BearerAuth\" followed by a space and Access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}